Removed seperated parse post functions in lib
